#110: Finish up auth pantheon yaml generation
